requirements.txt
# Python dependencies for Flow Manager

# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Async support
asyncio==3.4.3

# HTTP client (for potential external API calls)
httpx>=0.28.0

# Python dotenv for environment variables
python-dotenv>=1.0.0

# Rate limiting
slowapi==0.1.9

# Testing (optional but recommended)
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Code quality (optional)
black>=24.0.0
flake8>=7.0.0
pylint>=3.3.0
